// lib/content/types.ts
export interface FlashcardReviewExercise {
  type: "flashcard-review";
  front: string;
  back?: string;
  meaning?: string;
  translation?: string;
  pos?: string;
  ipa?: string;
  level?: string;
  cefrLevel?: string;
  noAudio?: string[];
  srsWords: string | string[];
}